fix(rstest): keep no-standalone-expect scope pushes and pops symmetric
The scope stack replaces upstream's enter/exit kind stack with source
ranges, which introduces a failure mode upstream does not have: a range
may be unavailable. Pushing only when one could be derived, while popping
for every call the enter handler considered, made a nested registration
without an inline callback close the enclosing test's scope, so every
assertion after `test.todo("x")`, `test("x", cb)` or a custom block
function in the same body was reported as standalone. The arrow branch
had the same asymmetry through its parent check.

Both sides now share one predicate and fall back to the node's own range.
That range also covers the arguments of a callback-less registration,
which brings `test.todo(expect(1).toBe(1))` in line with upstream.

The traversal moves to internal/utils/test_framework so the jest rule,
which carries the same defects in its own copy, can adopt it next.
